Store door flags on 16-bit and load 040 map

+++ src/mapdata/mmapper2exit.h
@@ -3,7 +3,7 @@
 ** Authors:   Ulf Hermann <ulfonk_mennhar@gmx.de> (Alve),
 **            Marek Krejza <krejza@gmail.com> (Caligor)
 **
-** This file is part of the MMapper project. 
+** This file is part of the MMapper project.
 ** Maintained by Nils Schimmelmann <nschimme@gmail.com>
 **
 ** This program is free software; you can redistribute it and/or
@@ -33,7 +33,7 @@ class Exit;
 
 enum ExitType { ET_NORMAL, ET_LOOP, ET_ONEWAY, ET_UNDEFINED };
 
-enum ExitDirection { ED_NORTH=0, ED_SOUTH, ED_EAST, ED_WEST, ED_UP, 
+enum ExitDirection { ED_NORTH=0, ED_SOUTH, ED_EAST, ED_WEST, ED_UP,
 			   ED_DOWN, ED_UNKNOWN, ED_NONE };
 
 ExitDirection opposite(ExitDirection in);
@@ -62,7 +62,7 @@ typedef class QString DoorName;
 enum ExitField {E_DOORNAME = 0, E_FLAGS, E_DOORFLAGS};
 
 typedef quint8 ExitFlags;
-typedef quint8 DoorFlags;
+typedef quint16 DoorFlags;
 
 ExitFlags getFlags(const Exit & e);
 
@@ -73,11 +73,11 @@ DoorFlags getDoorFlags(const Exit & e);
 void updateExit(Exit & e, ExitFlags flags);
 
 void orExitFlags(Exit & e, ExitFlags flags);
-  
+
 void nandExitFlags(Exit & e, ExitFlags flags);
 
 void orDoorFlags(Exit & e, DoorFlags flags);
-  
+
 void nandDoorFlags(Exit & e, DoorFlags flags);
-  
+
 #endif
